[reorg] Move AlternateSubst to new layout.
https://github.com/harfbuzz/harfbuzz/commit/6a369389d
