+++ src/hb/ot/layout/GSUB/mod.rs
@@ -1,3 +1,5 @@
+mod alternate_set;
+mod alternate_subst;
 mod multi_subst;
 mod sequence;
 mod single_subst;
