+++ src/hb/ot_layout_gsub_table.rs
@@ -1,5 +1,3 @@
-use core::convert::TryFrom;
-
 use ttf_parser::gsub::*;
 use ttf_parser::opentype_layout::LookupIndex;
 use ttf_parser::GlyphId;
@@ -9,54 +7,8 @@ use super::hb_font_t;
 use super::ot_layout::*;
 use super::ot_layout_common::{SubstLookup, SubstitutionTable};
 use super::ot_layout_gsubgpos::*;
-use super::ot_map::*;
 use super::ot_shape_plan::hb_ot_shape_plan_t;
 use OT::hb_ot_apply_context_t;
-
-impl Apply for AlternateSet<'_> {
-    fn apply(&self, ctx: &mut hb_ot_apply_context_t) -> Option<()> {
-        let len = self.alternates.len();
-        if len == 0 {
-            return None;
-        }
-
-        let glyph_mask = ctx.buffer.cur(0).mask;
-
-        // Note: This breaks badly if two features enabled this lookup together.
-        let shift = ctx.lookup_mask.trailing_zeros();
-        let mut alt_index = (ctx.lookup_mask & glyph_mask) >> shift;
-
-        // If alt_index is MAX_VALUE, randomize feature if it is the rand feature.
-        if alt_index == hb_ot_map_t::MAX_VALUE && ctx.random {
-            // Maybe we can do better than unsafe-to-break all; but since we are
-            // changing random state, it would be hard to track that.  Good 'nough.
-            ctx.buffer.unsafe_to_break(Some(0), Some(ctx.buffer.len));
-            alt_index = ctx.random_number() % u32::from(len) + 1;
-        }
-
-        let idx = u16::try_from(alt_index).ok()?.checked_sub(1)?;
-        ctx.replace_glyph(self.alternates.get(idx)?);
-
-        Some(())
-    }
-}
-
-// AlternateSubstFormat1::would_apply
-impl WouldApply for AlternateSubstitution<'_> {
-    fn would_apply(&self, ctx: &WouldApplyContext) -> bool {
-        ctx.glyphs.len() == 1 && self.coverage.get(ctx.glyphs[0]).is_some()
-    }
-}
-
-// AlternateSubstFormat1::apply
-impl Apply for AlternateSubstitution<'_> {
-    fn apply(&self, ctx: &mut hb_ot_apply_context_t) -> Option<()> {
-        let glyph = ctx.buffer.cur(0).as_glyph();
-        let index = self.coverage.get(glyph)?;
-        let set = self.alternate_sets.get(index)?;
-        set.apply(ctx)
-    }
-}
 
 impl WouldApply for Ligature<'_> {
     fn would_apply(&self, ctx: &WouldApplyContext) -> bool {
